% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orthologs_and_aliases.R
\name{get_orthologs_and_aliases}
\alias{get_orthologs_and_aliases}
\title{get_orthologs_and_aliases}
\usage{
get_orthologs_and_aliases(ref_species, list_of_interest)
}
\arguments{
\item{ref_species}{a string containing the nickname of a model species}

\item{list_of_interest}{a character vector containing gene symbols}
}
\value{
\strong{result}: a dataframe housing:
\itemize{
   \item the reference species
   \item the reference gene symbols
   \item the reference gene symbols' Ensembl IDs
   \item remaining (target) species names
   \item target species' orthologous gene symbols
   \item target species' orthologous gene Ensembl IDs
}
}
\description{
\strong{get_orthologs_and_aliases} is a function that finds all the aliases
   and orthologs of the other four main model organisms from a character
   vector of gene symbols of another
}
\details{
Queries the Bioconductor db to extract the aliases of the provided gene
   symbols, then uses the \strong{orthogene} package to extract orthologous
   gene information from the 5 main biological model species, and corrals
   into a practical and intuitive dataframe

\strong{ref_species} can be one of \strong{"human"}, \strong{"mouse"}, or
   \strong{"fly"}
}
\examples{
aDRG_DEG_list <- TMEM::aDRG_DEG_list
get_orthologs_and_aliases(ref_species = 'mouse',
                         list_of_interest = aDRG_DEG_list[c(1:5)])

}
\references{
[orthogene](https://github.com/neurogenomics/orthogene)
}
