% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMEM.R
\name{ortholog_and_alias_fn}
\alias{ortholog_and_alias_fn}
\title{ortholog_and_alias_fn}
\usage{
ortholog_and_alias_fn(ref_species, list_of_interest)
}
\arguments{
\item{ref_species}{a string containing the nickname of a model species}

\item{list_of_interest}{a character vector containing gene symbols}
}
\value{
\strong{`result`}: a dataframe housing:
 + the reference species
 + the reference gene symbols
 + the reference gene symbols' Ensembl IDs
 + remaining (target) species names
 + target species' orthologous gene symbols
 + target species' orthologous gene Ensembl IDs
}
\description{
\strong{`ortholog_and_alias_fn`} is a function that finds all the aliases and
orthologs of the other four main model organsims from a character vector of
gene symbols of another
}
\details{
Queries the Bioconductor db to extract the aliases of the provided gene symbols,
then uses the orthogene package to extract orthologous gene information from
the 5 main biological model species, and corrals into a practical and intuitive dataframe

\strong{`ref_species`} can be one of 'human', 'mouse', or 'fly'
}
\examples{
\dontrun{
ortholog_and_alias_fn(ref_species = 'human',
                     list_of_interest = my_character_vec)
                     }

}
\references{
[orthogene](https://github.com/neurogenomics/orthogene)
}
