% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMEM.R
\name{Find_Row_Z}
\alias{Find_Row_Z}
\title{Find_Row_Z}
\usage{
Find_Row_Z(Expression_Profile)
}
\arguments{
\item{Expression_Profile}{a dataframe containing a list of gene symbols (first column);
remaining columns comprising samples (iterations) of expression values}
}
\value{
\strong{`GeneZ`}: a dataframe with Z-scored expression values
}
\description{
\strong{`Find_Row_Z`} is a function that determines Z-scores of an expression matrix
}
\details{
Z-scores are \strong{`row-wise`} across all columns containing samples (column 2 through the end of the dataframe)
}
\examples{
\dontrun{
Find_Row_Z(Expression_Profile = matrix)
}
}
