% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orthologs_and_aliases.R
\name{get_orthologs_and_aliases}
\alias{get_orthologs_and_aliases}
\title{get_orthologs_and_aliases}
\usage{
get_orthologs_and_aliases(ref_species, list_of_interest)
}
\arguments{
\item{ref_species}{a string containing the nickname of a model species}

\item{list_of_interest}{a character vector containing gene symbols}
}
\value{
\strong{`result`}: a dataframe housing:
 + the reference species
 + the reference gene symbols
 + the reference gene symbols' Ensembl IDs
 + remaining (target) species names
 + target species' orthologous gene symbols
 + target species' orthologous gene Ensembl IDs
}
\description{
\strong{`get_orthologs_and_aliases`} is a function that finds all the aliases and
orthologs of the other four main model organisms from a character vector of
gene symbols of another
}
\details{
Queries the Bioconductor db to extract the aliases of the provided gene symbols,
then uses the `orthogene` package to extract orthologous gene information from
the 5 main biological model species, and corrals into a practical and intuitive dataframe

\strong{`ref_species`} can be one of 'human', 'mouse', or 'fly'
}
\examples{
\donttest{data("aDRG_DEG_list")
get_orthologs_and_aliases(ref_species = 'mouse',
                         list_of_interest = aDRG_DEG_list)
}

}
\references{
[orthogene](https://github.com/neurogenomics/orthogene)
}
